Create Hack_1_Hello_World
This hack will make Minecraft's chat pane display the text "Hello World". Once a successful test has been done, it might be cool to change the display text to the users name, or copy multiple instances of lines 5 & 6 to make Minecraft talk to you...maybe bet it to sing a song, tell a joke or do something entertaining.