Create Hack_2_Random_Messages.py
This code will display messages in Minecraft's Chat Pane. The messages it displays are stored in a list. The items contained in the list are 3 variables containing strings of data which display "Message 1" etc. These messages are indexed in the list at positions 0,1 & 2. We use the random.randint(0,2) command to choose a number at random between 0 and 2 (the aforementioned index of the strings in the list), and when this number is chosen we use it to select the message indexed in the list at this position. Users are encouraged to change the messages housed in the list, and to add more. In order to display these we will need to increase the range of the randint choice.