Create Hack_3_User_Messages.py
This hack utilises the Python shell to interact with the user. The use of the "While" loop ensures that our program runs forever. The shell will display the message "What should we say?". prompting the user to type a message. Once the user has typed the message and pressed enter, the input is saved as a string variable and displayed in the chat pane. Once the users message has been displayed in Minecraft the Python shell then resets and asks the question again.