Update Build-IT_World.Py
This is the original build file for our custom Minecraft world. It flattens the current world and builds a new one that we will use to develop the coding skills of young people. In order to run this, load it to the "mcpi" folder on your pi, and execute it with Python 2.7.
You will also need the logo.py file in the same location to run this.
NB: As it replaces over 15 million blocks it may take a couple of minutes to complete. Please be patient. It's worth it.